requirements.txt
python-telegram-bot==20.3
instaloader==4.10.1
python-dotenv==0.21.0
yt-dlp>=2023.10.13
